requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
cryptography
PyCryptodome
psycopg2
Flask-SocketIO
Flask-Migrate
requests
pycryptodome
python-socketio
requests
python-socketio
websocket-client
python-dotenv
apscheduler